feature/login: creating the chipList as a separate component
