+++ lib/src/auth/formFields/index.dart
@@ -2,3 +2,4 @@ export './customTextField.dart';
 export './customFormButton.dart';
 export './customImagePicker.dart';
 export './customDropdownField.dart';
+export './customChipList.dart';
